+++ lib/src/bones_api_entity_adapter_postgres.dart
@@ -34,12 +34,14 @@ class PostgreSQLAdapter extends SQLAdapter<PostgreSQLExecutionContext> {
   static FutureOr<PostgreSQLAdapter?> _instantiate(config,
       {int? minConnections,
       int? maxConnections,
-      EntityRepositoryProvider? parentRepositoryProvider}) {
+      EntityRepositoryProvider? parentRepositoryProvider,
+      String? workingPath}) {
     try {
       return PostgreSQLAdapter.fromConfig(config,
           minConnections: minConnections,
           maxConnections: maxConnections,
-          parentRepositoryProvider: parentRepositoryProvider);
+          parentRepositoryProvider: parentRepositoryProvider,
+          workingPath: workingPath);
     } catch (e, s) {
       _log.severe("Error instantiating from config", e, s);
       return null;
@@ -65,7 +67,8 @@ class PostgreSQLAdapter extends SQLAdapter<PostgreSQLExecutionContext> {
       bool generateTables = false,
       Object? populateTables,
       Object? populateSource,
-      EntityRepositoryProvider? parentRepositoryProvider})
+      EntityRepositoryProvider? parentRepositoryProvider,
+      String? workingPath})
       : host = host ?? 'localhost',
         port = port ?? 5432,
         _password = (password != null && password is! PasswordProvider
@@ -85,6 +88,7 @@ class PostgreSQLAdapter extends SQLAdapter<PostgreSQLExecutionContext> {
           populateTables: populateTables,
           populateSource: populateSource,
           parentRepositoryProvider: parentRepositoryProvider,
+          workingPath: workingPath,
         ) {
     boot();
 
@@ -102,7 +106,8 @@ class PostgreSQLAdapter extends SQLAdapter<PostgreSQLExecutionContext> {
       int? defaultPort,
       int? minConnections,
       int? maxConnections,
-      EntityRepositoryProvider? parentRepositoryProvider}) {
+      EntityRepositoryProvider? parentRepositoryProvider,
+      String? workingPath}) {
     boot();
 
     var host = config?['host'] ?? defaultHost;
@@ -142,6 +147,7 @@ class PostgreSQLAdapter extends SQLAdapter<PostgreSQLExecutionContext> {
       populateTables: populateTables,
       populateSource: populateSource,
       parentRepositoryProvider: parentRepositoryProvider,
+      workingPath: workingPath,
     );
   }